move ship factories to /factories directory
