+++ src/factories/ship.js
